smt/Number: use longs for numerical values
The minimum integer is larger than the max integer if the sign is
removed, which results in an exception when attempting to parse.
Perhaps not the most efficient solution, since we know the actual output
is always within the bounds of an integer, we can simply use the long
type instead.

+++ src/main/java/processing/smt/Number.java
@@ -8,15 +8,15 @@ import soot.Value;
 import soot.jimple.IntConstant;
 
 public class Number extends SmtExpression {
-    private final int value;
+    private final long value;
 
-    public Number(int value) {
+    public Number(long value) {
         super();
         this.value = value;
     }
 
     public Value getValue() {
-        return IntConstant.v(this.value);
+        return IntConstant.v((int)this.value);
     }
 
     public Optional<Value> getValue(Set<Local> variables) {
